cisstCommon, fixed JSON for Windows 64 bits

+++ cisstCommon/code/cmnDataFunctionsJSON.cpp
@@ -74,11 +74,11 @@ void cmnDataJSON<unsigned int>::DeSerializeText(DataType & data, const Json::Val
 }
 
 template <>
-void cmnDataJSON<unsigned long int>::SerializeText(const DataType & data, Json::Value & jsonValue) {
+void cmnDataJSON<unsigned long long int>::SerializeText(const DataType & data, Json::Value & jsonValue) {
     jsonValue = Json::UInt64(data);
 }
 template <>
-void cmnDataJSON<unsigned long int>::DeSerializeText(DataType & data, const Json::Value & jsonValue) throw (std::runtime_error) {
+void cmnDataJSON<unsigned long long int>::DeSerializeText(DataType & data, const Json::Value & jsonValue) throw (std::runtime_error) {
     data = jsonValue.asUInt64();
 }
